Description: Scans the LAN network and returns the IPs and MAC address of the devices connected to the same network. - added function networkdetails that is getting the ip addr and netmask used by the device (Windows only) - modified the networkdetails so that will convert the subnetmask to cird (example from 255.255.255.0 to /28). The network IP together with cird will be passed to scanet function. - input for the postfix was hard coded, and modified it so that it will take the network address / subnetmask - added expTopJson, that will write the dictionary with IPs and Mac addresses in a json file - modified how the dictionary is created, the element is of type list. - modified the scannet function so that will create a python dictionary where the key is the MAC address and the values is the IP - created function expToJson that will get the dictionary returned by the scanned function and will write in a file, on the 1st run. If the file exists the function will keep track on how many times that mac address was found being connected to the same LAN as you.

+++ expToJson.py
@@ -4,6 +4,28 @@ import os
 def export_to_json_file(dict_in):
     jfile_name = "jdb.json"
     file_dict = {}
-    with open(jfile_name, "w") as json_file:
+
+    if os.path.exists(jfile_name) and os.path.getsize(jfile_name) > 0:
+        with open(jfile_name, "r") as json_file:
+            jdict = json.load(json_file)
+        print("this is {}".format(jdict))
+
+        new_dict = jdict.copy()
+        for j, k in dict_in.items():
+
+            if j in jdict:
+
+                new_dict[j] = new_dict[j] + 1
+            else:
+                new_dict[j] = 1
+
+        print("new: {}".format(new_dict))
+
+        with open(jfile_name, "w") as json_file:
+            json.dump(new_dict, json_file)
+
+
+    else:
+        with open(jfile_name, "w") as json_file:
             json.dump(dict_in, json_file)
